Create a permission_error page for permission related messages
Related to issue #22, instead of logging of the users anytime
there is a permission issue, now the user will be redirected
to a permission_error page that does not log him off.

If the user was not logged in in the first place, the behavior
is not changed, the user will be redirected to the login page
with a relevant message instead.